chore: move schemas in strategies folder
