--- docs/api/schemas.rst
@@ -1,6 +1,0 @@
-Schemas
-========
-
-
-.. automodule:: substrafl.schemas
-    :private-members:
