+++ substrafl/__init__.py
@@ -7,7 +7,6 @@ from substrafl.index_generator.np_index_generator import NpIndexGenerator
 from substrafl.logger import set_logging_level
 from substrafl.model_loading import download_algo_files
 from substrafl.model_loading import load_algo
-from substrafl.schemas import StrategyName
 
 set_logging_level(loglevel=logging.INFO)
 
@@ -16,7 +15,6 @@ __all__ = [
     "set_logging_level",
     "NpIndexGenerator",
     "EvaluationStrategy",
-    "StrategyName",
     "load_algo",
     "download_algo_files",
 ]
