+++ substrafl/strategies/__init__.py
@@ -2,7 +2,8 @@ from substrafl.strategies.fed_avg import FedAvg
 from substrafl.strategies.fed_pca import FedPCA
 from substrafl.strategies.newton_raphson import NewtonRaphson
 from substrafl.strategies.scaffold import Scaffold
+from substrafl.strategies.schemas import StrategyName
 from substrafl.strategies.single_organization import SingleOrganization
 from substrafl.strategies.strategy import Strategy
 
-__all__ = ["Strategy", "FedAvg", "FedPCA", "SingleOrganization", "Scaffold", "NewtonRaphson"]
+__all__ = ["Strategy", "StrategyName", "FedAvg", "FedPCA", "SingleOrganization", "Scaffold", "NewtonRaphson"]
