+++ substrafl/strategies/schemas.py
