+++ tests/conftest.py
@@ -17,7 +17,7 @@ from substrafl.nodes.aggregation_node import AggregationNode
 from substrafl.nodes.test_data_node import TestDataNode
 from substrafl.nodes.train_data_node import TrainDataNode
 from substrafl.remote.decorators import remote_data
-from substrafl.schemas import StrategyName
+from substrafl.strategies.schemas import StrategyName
 from substrafl.strategies.strategy import Strategy
 
 from . import assets_factory
